requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-WTF==1.2.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
paypalrestsdk==1.13.3
pycparser==2.21
pyOpenSSL==24.0.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
stripe==8.4.0
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
